+++ src/Billing.WebApp/Entities/InvoiceItem.cs
@@ -10,7 +10,7 @@ namespace Billing.WebApp.Entities
         public int Quantity { get; set; }
         public float Price { get; set; }
         public string Description { get; set; }
-        public int TaxAmount { get; set; }
+        public float TaxAmount { get; set; }
         public bool TaxPercentage { get; set; } = true;
         public bool TaxInclusive { get; set; } = true;
         public DateTime Created { get; set; } = DateTime.Now;
